refactor(subcommands): separating sub-commands into separate files
