+++ engine/command/blockchain/blockchain.go
@@ -1,13 +1,8 @@
 package blockchain
 
 import (
-	"fmt"
-	"strconv"
-
-	"github.com/pactus-project/pactus/types/amount"
 	"github.com/pagu-project/Pagu/client"
 	"github.com/pagu-project/Pagu/engine/command"
-	"github.com/pagu-project/Pagu/utils"
 )
 
 const (
@@ -81,58 +76,4 @@ func (bc *Blockchain) GetCommand() command.Command {
 	cmdBlockchain.AddSubCommand(subCmdCalcFee)
 
 	return cmdBlockchain
-}
-
-func (bc *Blockchain) calcRewardHandler(cmd command.Command, _ command.AppID, _ string, args ...string) command.CommandResult {
-	stake, err := strconv.Atoi(args[0])
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	time := args[1]
-
-	if stake < 1 || stake > 1_000 {
-		return cmd.ErrorResult(fmt.Errorf("%v is invalid amount; minimum stake amount is 1 PAC and maximum is 1,000 PAC", stake))
-	}
-
-	var blocks int
-	switch time {
-	case "day":
-		blocks = 8640
-	case "month":
-		blocks = 259200
-	case "year":
-		blocks = 3110400
-	default:
-		blocks = 8640
-		time = "day"
-	}
-
-	bi, err := bc.clientMgr.GetBlockchainInfo()
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	reward := int64(stake*blocks) / int64(amount.Amount(bi.TotalPower).ToPAC())
-
-	return cmd.SuccessfulResult("Approximately you earn %v PAC reward, with %v PAC stake 🔒 on your validator in one %s ⏰ with %s total power ⚡ of committee."+
-		"\n\n> Note📝: This number is just an estimation. It will vary depending on your stake amount and total network power.",
-		utils.FormatNumber(reward), utils.FormatNumber(int64(stake)), time, utils.FormatNumber(int64(amount.Amount(bi.TotalPower).ToPAC())))
-}
-
-func (bc *Blockchain) calcFeeHandler(cmd command.Command, _ command.AppID, _ string, args ...string) command.CommandResult {
-	amt, err := amount.FromString(args[0])
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	fee, err := bc.clientMgr.GetFee(int64(amt))
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	calcedFee := amount.Amount(fee)
-
-	return cmd.SuccessfulResult("Sending %s will cost %s with current fee percentage."+
-		"\n> Note: Consider unbond and sortition transaction fee is 0 PAC always.", amt, calcedFee.String())
 }
