+++ engine/command/phoenix/faucet.go
@@ -1,0 +1,43 @@
+package phoenix
+
+import (
+	"github.com/pagu-project/Pagu/database"
+	"github.com/pagu-project/Pagu/engine/command"
+)
+
+func (pt *Phoenix) faucetHandler(cmd command.Command, _ command.AppID, callerID string, args ...string) command.CommandResult {
+	if !pt.db.HasUser(callerID) {
+		if err := pt.db.AddUser(
+			&database.User{
+				ID: callerID,
+			},
+		); err != nil {
+			return cmd.ErrorResult(err)
+		}
+	}
+
+	if !pt.db.CanGetFaucet(callerID) {
+		return cmd.FailedResult("Uh, you used your share of faucets today!")
+	}
+
+	if pt.wallet.Balance() < 5 {
+		return cmd.FailedResult("RoboPac Phoenix wallet is empty, please contact the team!")
+	}
+
+	toAddr := args[0]
+	txID, err := pt.wallet.TransferTransaction(toAddr, "Phoenix Testnet Pagu Faucet", 5) //! define me on config?
+	if err != nil {
+		return cmd.ErrorResult(err)
+	}
+
+	if err = pt.db.AddFaucet(&database.Faucet{
+		Address:         toAddr,
+		Amount:          5,
+		TransactionHash: txID,
+		UserID:          callerID,
+	}); err != nil {
+		return cmd.ErrorResult(err)
+	}
+
+	return cmd.SuccessfulResult("You got %d tPAC in %s address on Phoenix Testnet!", 5, toAddr)
+}
