+++ engine/command/zealy/zealy.go
@@ -69,61 +69,3 @@ func (z *Zealy) GetCommand() command.Command {
 
 	return cmdZealy
 }
-
-func (z *Zealy) claimHandler(cmd command.Command, _ command.AppID, callerID string, args ...string) command.CommandResult {
-	user, err := z.db.GetZealyUser(callerID)
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	if user.IsClaimed {
-		return cmd.FailedResult("You already claimed your reward: https://pacviewer.com/transaction/%s",
-			user.TxHash)
-	}
-
-	address := args[0]
-	txHash, err := z.wallet.TransferTransaction(address, "Pagu Zealy reward distribution", int64(user.Amount))
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	if err = z.db.UpdateZealyUser(callerID, txHash); err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	return cmd.SuccessfulResult("Zealy reward claimed successfully: https://pacviewer.com/transaction/%s",
-		txHash)
-}
-
-func (z *Zealy) statusHandler(cmd command.Command, _ command.AppID, _ string, args ...string) command.CommandResult {
-	allUsers, err := z.db.GetAllZealyUser()
-	if err != nil {
-		return cmd.ErrorResult(err)
-	}
-
-	total := 0
-	totalClaimed := 0
-	totalNotClaimed := 0
-
-	totalAmount := 0
-	totalClaimedAmount := 0
-	totalNotClaimedAmount := 0
-
-	for _, u := range allUsers {
-		total++
-		totalAmount += int(u.Amount)
-
-		if u.IsClaimed {
-			totalClaimed++
-			totalClaimedAmount += int(u.Amount)
-		} else {
-			totalNotClaimed++
-			totalNotClaimedAmount += int(u.Amount)
-		}
-	}
-
-	return cmd.SuccessfulResult("Total Users: %v\nTotal Claims: %v\nTotal not remained claims: %v\nTotal Coins: %v PAC\n"+
-		"Total claimed coins: %v PAC\nTotal not claimed coins: %v PAC\n", total, totalClaimed, totalNotClaimed,
-		totalAmount, totalClaimedAmount, totalNotClaimedAmount,
-	)
-}
